Enhance getUserLocale to better find the best lang
The current implementation of getUserLocale didn't catch regional
variations of the languages. These are separated by a dash and uppercase
letters. E.g. a user with 'de-CH' didn't get the German translation but
fell back to English.

This new implementation also considers these variations and falls back
to the more general language if the exact variation is not available.

Additionally, instead of navigator.language we now use
navigator.languages, which contains all of the browsers preferred
languages in order. This features is supported by all major browsers
since 2017. ['it', 'de'] will then resolve to German instead of English.